spliting OreidContext to its own file
